% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affichage_general.R
\name{affichage_general}
\alias{affichage_general}
\title{Affichage matrices boxplots et histogrammes}
\usage{
affichage_general(
  Z,
  r,
  titre = "",
  paletteinf = "",
  palettesup = "",
  mmechelle = FALSE
)
}
\arguments{
\item{Z}{matrice}

\item{r}{rayon de lissage}

\item{titre}{titre du graphique}

\item{paletteinf}{to fix color on the scale}

\item{palettesup}{to fix color on the scale}

\item{mmechelle}{to set the two same scales TRUE/FALSE}
}
\value{
a plot
}
\description{
Affichage matrices boxplots et histogrammes
}
\examples{
affichage_general(matrix(rnorm(100), nrow = 10), r =2)
}
