\name{Changings}
\alias{Changings}
\alias{changings}

\alias{tbmdim}
\alias{coord_units}
\alias{new_coord_units}
\alias{variab_units}

\title{Documentation of some further changings} 

\description{
  \itemize{
    \item{Options getting obsolete}
   \itemize{
     \item\code{oldstyle} is becoming \code{warn_oldstyle} 
     \item\code{newstyle} is becoming \code{warn_newstyke} 
     \item\code{newAniso} is becoming \code{warn_newAniso} 
     \item\code{ambiguous} is becoming \code{warn_ambiguous} 
     \item\code{normal_mode} is becoming \code{warn_normal_mode} 
     \item\code{colour_palette} is becoming \code{warn_colour_palette} 
   }

    
   \item \bold{Changings in option names}
   \itemize{
     \item several changes in \code{RFoptions()$graphics} in version 3.1.11
     \item\code{pdfnumber} became in version 3.0.42 \code{filenumber}      
     \item\code{pdfonefile} became in version 3.0.42 \code{onefile}      
     \item\code{pdffile} became in version 3.0.42 \code{file}      
     \item\code{tbmdim} became in version 3.0.41 \code{reduceddim} 
     \item\code{coord_units} became in version 3.0.39 \code{coordunits} 
     \item\code{new_coord_units} became in version 3.0.39 \code{new_coordunits} 
     \item\code{variab_units} became in version 3.0.39 \code{varunits}      
    }

%   \item \bold{Changings in argument names}
%   \itemize{
%     \item 
%   }

 } 
}

\seealso{
  \link{MajorRevisions},
  \link{RandomFields}
}

\examples{
## no examples given
\dontshow{FinalizeExample()}
}
