% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charactervector_as_vector.R
\name{charactervector_as_vector}
\alias{charactervector_as_vector}
\title{charactervector_as_vector}
\usage{
charactervector_as_vector(chaine)
}
\arguments{
\item{chaine}{Ex : x <- "(0, 1)" a string}
}
\value{
a vector Ex : c(0, 1)
}
\description{
charactervector_as_vector
}
\examples{
charactervector_as_vector("(0, 1)")

}
