\name{CheckXYZ}
\alias{CheckXYZ}
\title{Internal function -- do not use it directly}
\description{
  checks whether certain parameters are specified correctly
}
\usage{
CheckXYZ(x, y, z, grid, gridtriple)
}
\arguments{
  \item{x}{\code{x} coordinates}
  \item{y}{\code{y} coordinates}
  \item{z}{\code{z} coordinates}
  \item{grid}{see \code{\link{GaussRF}}
  \item{gridtriple}{see \code{\link{GaussRF}}
}
\value{
  list of internal parameters
}
\author{Martin Schlather, \email{Martin.Schlather@uni-bayreuth.de}
  \url{http://www.geo.uni-bayreuth.de/~martin}}
\keyword{ spatial }%-- one or more ...
