% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_actual_covariance.R
\name{data_frame_actual_covariance}
\alias{data_frame_actual_covariance}
\title{data_frame_actual_covariance}
\usage{
data_frame_actual_covariance(
  Z,
  les_rayons,
  les_directions,
  echelle = 1,
  vecteur_directeur = TRUE
)
}
\arguments{
\item{Z}{the matrix}

\item{les_rayons}{radius of moving average window in a vector}

\item{les_directions}{list of direction vectors}

\item{echelle}{scale (optional)}

\item{vecteur_directeur}{set the direction with minimum numbers in the vector}
}
\value{
a dtaframe Distance_pixels Distance_km Empirical_covariance Empirical_correlation
}
\description{
data_frame_actual_covariance
}
\examples{
data_frame_actual_covariance(matrix(rnorm(1:100), nrow = 10), c(1,2), list(c(0,1), c(1,1)))
}
