% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_expected_cov.R
\name{data_frame_expected_cov}
\alias{data_frame_expected_cov}
\title{Dataframe for the expected covariance}
\usage{
data_frame_expected_cov(
  sigma_Z,
  the_radius,
  the_directions,
  the_scale = 1,
  direction = TRUE
)
}
\arguments{
\item{sigma_Z}{the standard error of the theoritical distribution}

\item{the_radius}{the radius of the window of the moving average}

\item{the_directions}{the directions we consider in a list ex : list(c(0,1), c(1, 1))}

\item{the_scale}{how many km per pixels ?}

\item{direction}{TRUE by default. reduce the direction into its director vector}
}
\value{
A dataframe with seven colomns : the distance between two points in pixels unit, this distance in km unit,
the value of the expected covariance and expected correlation
the considered direction, and the radius of the window of the moving average (in pixels and then km unit).
}
\description{
Dataframe for the expected covariance
}
\examples{
data_frame_expected_cov(1, c(1), list(c(0,1), c(1,1), c(1, 2)))

}
