% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_covariance.R
\name{expected_covariance}
\alias{expected_covariance}
\title{Expected covariance}
\usage{
expected_covariance(sigma_Z, r, the_direction, the_scale = 1)
}
\arguments{
\item{sigma_Z}{the standard error of the theoritical distribution}

\item{r}{the radius of the moving average window}

\item{the_direction}{the direction we consider ex : c(1, 1)}

\item{the_scale}{how many km/pixels}
}
\value{
the covariance according to the distance for a given direction
}
\description{
Expected covariance
}
\examples{
expected_covariance(1, 1, c(0, 1))


}
