\name{Forest}
\alias{Forest}
\title{Simulation study for Forest data}
\description{
  -- to be done -- currently for internal use only
}
\usage{
Forest(x, model, percent, ref.edge, coarse, gausspercent, debug=FALSE)
}
\arguments{
  \item{x}{--}
  \item{model}{--}
  \item{percent}{--}
  \item{ref.edge}{--}
  \item{coarse}{--}
  \item{gausspercent}{--}
  \item{debug}{--}
}
\value{
  --
}
\details{
  --
}
\references{
  --
}
%\examples{
%}
\author{Martin Schlather, \email{martin.schlather@math.uni-goettingen.de}
  \url{http://www.stochastik.math.uni-goettingen.de/~schlather}
}
\keyword{spatial}
