% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iswholenumber.R
\name{iswholenumber}
\alias{iswholenumber}
\title{iswholenumber}
\usage{
iswholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{the number to test : is it an integer ?}

\item{tol}{tolerance}
}
\value{
TRUE/FALSE : test if x is an integer
}
\description{
iswholenumber
}
\examples{
iswholenumber(22)
}
