% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_boxplots.R
\name{les_boxplots}
\alias{les_boxplots}
\title{Boxplots}
\usage{
les_boxplots(Z, Y, titre = "Boxplots", nom_axeX = "", nom_axeY = "")
}
\arguments{
\item{Z}{first matrix}

\item{Y}{second matrix}

\item{titre}{title of the graph}

\item{nom_axeX}{x label}

\item{nom_axeY}{y label}
}
\value{
a plot with the boxplots
}
\description{
Boxplots
}
\examples{
les_boxplots(1:10, 10:20)
}
