% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_histogrammes.R
\name{les_histogrammes}
\alias{les_histogrammes}
\title{lLs histogrammes}
\usage{
les_histogrammes(Z, Y, titre = "Histogrammes", nom_axeX = "", nom_axeY = "")
}
\arguments{
\item{Z}{a matrix}

\item{Y}{another matrix}

\item{titre}{title if histograms}

\item{nom_axeX}{x label}

\item{nom_axeY}{y label}
}
\value{
a plot with 2 histograms
}
\description{
lLs histogrammes
}
\examples{
les_histogrammes(matrix(1:100, nrow = 10), matrix(101:200, nrow = 10))
}
