% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liste.R
\name{liste}
\alias{liste}
\title{Liste}
\usage{
liste(v)
}
\arguments{
\item{v}{a vector with many strings of vectors Ex :  c("(0, 1)", "(1, 1)")}
}
\value{
a list of vectors
}
\description{
Liste
}
\examples{
liste(c("(0, 1)", "(1, 1)"))
}
