% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_shaped_matrix.R
\name{long_shaped_matrix}
\alias{long_shaped_matrix}
\title{Long shaped matrix}
\usage{
long_shaped_matrix(Z)
}
\arguments{
\item{Z}{a matrix/dataframe/vector}
}
\value{
a dataframe : first column is x coordinates, second column is y coordinate, third column is the value
}
\description{
Long shaped matrix
}
\examples{
Z <- matrix(1:25, nrow = 5)
long_shaped_matrix(Z)
}
