% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving_average}
\alias{moving_average}
\title{Moving average}
\usage{
moving_average(Z, r)
}
\arguments{
\item{Z}{matrix/dataframe}

\item{r}{integer : the radius of the window of the moving average. Ex : r = 1 implies a window of 3*3}
}
\value{
matrix (side effects imply a smaller matrix than the Z matrix passed in parametre)
}
\description{
Create a new map/matrix by doing a moving average on a matrix
}
\examples{
moving_average(matrix(rnorm(10*10, 0,1), nrow = 10), r=2)
}
