% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expected_cov.R
\name{plot_expected_cov}
\alias{plot_expected_cov}
\title{Title}
\usage{
plot_expected_cov(
  sigma_Z,
  the_radius,
  the_directions,
  xlabs = "Distance (pixel unit)",
  ylabs = "Expected covariance",
  x = "Distance_pixels",
  y = "Expected_covariance",
  rayon = "Radius_pixels",
  the_scale = 1,
  max = "",
  maxy_sup = "",
  maxy_inf = "",
  connect = TRUE,
  director_vector = TRUE
)
}
\arguments{
\item{sigma_Z}{the standard error}

\item{the_radius}{radius of the window of the moving average (many are possibles ex : c(1, 2, 3))}

\item{the_directions}{ex : list(c(0,1), c(1,1))}

\item{xlabs}{the label of x axis}

\item{ylabs}{the label of y axis}

\item{x}{A string : "Distance_pixels" by default, or "Distance_km"}

\item{y}{A string : ""Expected covariance"" by default, or "Expected correlation}

\item{rayon}{A string for the unit of radius : "Radius_pixels" by default, or "Radius_km"}

\item{the_scale}{how many km/pixel}

\item{max}{the maximum value for x axis (the minimum is always zero)}

\item{maxy_sup}{the maximum value for y axis}

\item{maxy_inf}{the minimum value for y axis}

\item{connect}{TRUE by default : to connect points for a given radius and direction}

\item{director_vector}{TRUE by default : the given direction is transformed : it takes the director vector}
}
\value{
a plot of the expected covariance/correlation according to the distance
}
\description{
Title
}
\examples{
plot_expected_cov(1, c(1,2,3), list(c(0,1), c(1,1), c(1,2)))
}
