% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot a matrix}
\usage{
plot_matrix(
  M,
  r = "",
  paletteinf = "",
  palettesup = "",
  titre = "",
  nom_axeX = "",
  nom_axeY = "",
  echelle = "Scale"
)
}
\arguments{
\item{M}{a matrix/vector/number/dataframe of numbers}

\item{r}{a radius for the window of the moving average (optional)}

\item{paletteinf}{extreme color of color scale}

\item{palettesup}{extreme color of color scale}

\item{titre}{a title for the matrix}

\item{nom_axeX}{name for X axis}

\item{nom_axeY}{name for Y axis}

\item{echelle}{title of scale}
}
\value{
a plot
}
\description{
Plot a matrix
}
\examples{
plot_matrix(matrix(rbinom(10*10, 1, 0.5), nrow = 10))
plot_matrix(matrix(rbinom(10*10, 1, 0.5), nrow = 10), r = 2)
}
