\name{sleep.milli}
\alias{sleep.milli}
\alias{sleep}
\title{Sleep}
\description{
  Process sleeps for a given amount of time
}
\usage{
sleep.milli(milli)
}
\arguments{
  \item{milli}{sleeping time in milliseconds}
}
\value{
  No value is returned.
}
\author{Martin Schlather, \email{schlath@hsu-hh.de}
  \url{http://www.unibw-hamburg.de/WWEB/math/schlath/schlather.html}}
\keyword{utilities}
\keyword{misc}
