% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{variance}
\alias{variance}
\title{Variance}
\usage{
variance(M)
}
\arguments{
\item{M}{a number, a vector or a matrix, with or without NA}
}
\value{
a float (variance )
}
\description{
Returns the biased variance (division by n, not by (n-1))
}
\examples{
variance(c(1, 2, 3))
variance (matrix(rbinom(2*2, 1, 0.5), nrow = 2))

}
