\name{Baysian}
\alias{baysian}
\alias{Baysian}
\title{Baysian Spatial Modelling}
\description{
  Baysian modelling is a future project.
}
%\usage{}
%\arguments{}
%\value{}
\details{
  -- to be done --
}
%\references{Ribeiro}
\seealso{
  \link{RMmodelsAdvanced}
}

\examples{
RFoptions(seed=0)

# to do
\dontshow{if (.C("isAuthor", a=integer(1))$a) {
}}

\dontshow{RFoptions(seed=NA)}

}
\author{Martin Schlather, \email{schlather@math.uni-mannheim.de}
}
\keyword{spatial}
