\name{Baysian}
\alias{baysian}
\alias{Baysian}
\title{Baysian Spatial Modelling}
\description{
  Baysian modelling is a future project. However, some kind of
  hierarchical modelling is already available, see \link{RRmodels}.
}
%\usage{}
%\arguments{}
%\value{}
\details{
  -- to be done --
}
%\references{Ribeiro}
\seealso{
  \link{RMmodelsAdvanced}
}

\examples{
RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again


# to do

\dontshow{FinalizeExample()}

}
\author{Martin Schlather, \email{schlather@math.uni-mannheim.de}
}
\keyword{spatial}
