\name{papers}
\alias{papers}
%to do: check all
\title{Papers involving \pkg{RandomFields} and co-authored by M. Schlather}
\description{
 Here, an overview is given over the papers co-authored by M. Schlather
 that involve \pkg{RandomFields}.
 }
 
 \author{Martin Schlather, \email{schlather@math.uni-mannheim.de}
 \url{http://ms.math.uni-mannheim.de/de/publications/software};
 }
 \references{
 \itemize{
 \item
 Gneiting, T., Kleiber, W. and Schlather, M. (2010) Matern
 cross-covariance functions for multivariate random fields 
 \emph{J. Amer. Statist. Assoc.} \bold{105}, 1167-1177.

 See \link{GKS11} for the code.
 
 \item
 Gneiting, T., Sevcikova, H., Percival, D.B., Schlather, M., Jiang,
 Y. (2006) Fast and Exact Simulation of Large Gaussian Lattice Systems
 in R2: Exploring the Limits. \emph{J. Comput. Graph. Stat.},
 \bold{15}, 483-501.

 See \link{GSPSJ06} for the code.

 \item
 Scheuerer, M. and Schlather, M. (2012) Covariance Models for Random
 Vector Fields. \emph{Stochastic Models}, \bold{82}, 433-451.

 See \link{SS12} for the code. 

 \item
 Schlather, M. (2002) Models for stationary max-stable
 random fields. \emph{Extremes} \bold{5}, 33-44.

 See \link{S02} for the code.

 
 \item
 Schlather, M. (2010)
 On some covariance models based on normal scale mixtures.
 \emph{Bernoulli}, \bold{16}, 780-797.

 See \link{S10} for the code.


 \item
 Schlather, M., Malinowski, A., Menck, P.J., Oesting, M. and Strokorb, K. (2014)
 Analysis, simulation and prediction of
 multivariate random fields with package RandomFields \emph{JSS}
 \emph{Submitted}

 See \link{jss14} for the code.
 

 \item
 Strokorb, K., Ballani, F. and Schlather, M. (2014)
 In Preparation.

 See \link{SBS14} for the code.
 
}
 % Schlather, M. (2001) Simulation of stationary and isotropic random
% fields. \emph{R-News} \bold{1} (2), 18-20.


}

\seealso{
 \link{weather}, \link{GSPSJ06}, \link{SS12}, \link{S02}, \link{S10}, \link{jss14}
 }


\examples{
# for examples, see the specific papers.
\dontshow{FinalizeExample()}

}

\keyword{spatial}


