% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actual_cov.R
\name{plot_actual_cov}
\alias{plot_actual_cov}
\title{plot_actual_covariance}
\usage{
plot_actual_cov(
  Z,
  les_rayons,
  les_directions,
  ylabs = "",
  xlabs = "Distance between variables",
  x = "Distance",
  y = "Empirical_covariance",
  rayon = "Radius",
  echelle = 1,
  max = "",
  maxy_sup = "",
  maxy_inf = "",
  relier = TRUE,
  vecteur_directeur = TRUE
)
}
\arguments{
\item{Z}{a matrix}

\item{les_rayons}{radius of the moving average (vector)}

\item{les_directions}{directions (list)}

\item{ylabs}{set y legend}

\item{xlabs}{set x legend}

\item{x}{x parametre}

\item{y}{y parametre}

\item{rayon}{"Radius" or "Rayon_km" (radius unit)}

\item{echelle}{scale}

\item{max}{set x max}

\item{maxy_sup}{set y max}

\item{maxy_inf}{set y min}

\item{relier}{connect points}

\item{vecteur_directeur}{TRUE/FALSE : ex : c(2,4) become c(1,2)}
}
\value{
a plot
}
\description{
plot_actual_covariance
}
\examples{
plot_actual_cov(matrix(rnorm(10000, 0, 1), nrow = 100), c(1,2), list(c(0,1), c(1,1)), max = 20)
}
