\name{sleep.milli}
\alias{sleep.milli}
\alias{sleep}
\title{Sleep}
\description{
  Process sleeps for a given amount of time
}
\usage{
sleep.milli(milli)
}
\arguments{
  \item{milli}{sleeping time in milliseconds}
}
\value{
  No value is returned.
}
\author{Martin Schlather, \email{martin.schlather@math.uni-goettingen.de}
  \url{http://www.stochastik.math.uni-goettingen.de/~schlather}}
\keyword{utilities}
\keyword{misc}
